+++ pkg/vault/storage/etcd/etcd_test.go
@@ -1,4 +1,4 @@
-package util
+package etcd
 
 import (
 	"testing"
@@ -64,21 +64,12 @@ tls-key-file = "/etc/vault/storage/etcd/tls/etcd-client.key"
 
 	for _, test := range testaData {
 		t.Run(test.testName, func(t *testing.T) {
-			config, err := GetEtcdConfig(test.etcdSpec)
+			etcd, err := NewOptions(*test.etcdSpec)
+			assert.Nil(t, err)
+
+			config, err := etcd.GetStorageConfig()
 			assert.Nil(t, err)
 			assert.Equal(t, test.expectedOutput, config)
 		})
 	}
-}
-
-func TestGetListenerConfig(t *testing.T) {
-	expectedOutput := `
-listener "tcp" {
-  address = "0.0.0.0:8200"
-  cluster_address = "0.0.0.0:8201"
-  tls_cert_file = "/etc/vault/tls/server.crt"
-  tls_key_file  = "/etc/vault/tls/server.key"
-}
-`
-	assert.Equal(t, expectedOutput, GetListenerConfig())
 }
